moved Locomotion function to its own file in Controllers directory
