+++ projectApp/src/App.tsx
@@ -1,17 +1,11 @@
 import "./styles/App.css";
-import { Canvas, useFrame } from "@react-three/fiber";
+import { Canvas } from "@react-three/fiber";
 import InitScene from "./pages/Scene/InitScene";
 import { Sky } from "@react-three/drei";
-import {
-  createXRStore,
-  useXR,
-  useXRInputSourceState,
-  XR,
-  XROrigin,
-} from "@react-three/xr";
+import { createXRStore, useXR, XR } from "@react-three/xr";
 import BrowserUI from "./pages/UI/BrowserUI";
-import { useRef, useState } from "react";
-import { Group } from "three";
+import { useState } from "react";
+import Locomotion from "./controller/Locomotion";
 
 //Initializes and configures various parts integral to a VR experienceq
 const store = createXRStore();
@@ -63,36 +57,4 @@ function App() {
   );
 }
 
-function Locomotion({ speed }: { speed: number }) {
-  const controller = useXRInputSourceState("controller", "right");
-  const ref = useRef<Group>(null);
-  useFrame((_, delta) => {
-    if (ref.current == null || controller == null) {
-      return;
-    }
-    const thumstickState = controller.gamepad["xr-standard-thumbstick"];
-    if (thumstickState == null) {
-      return;
-    }
-    const headset = ref.current.children[0]; // Assuming the first child is the camera
-
-    // Apply headset rotation (approximating quaternion application without `three.js`)
-    const angle = Math.atan2(
-      -headset.matrix.elements[8],
-      headset.matrix.elements[10],
-    );
-    const cos = Math.cos(angle);
-    const sin = Math.sin(angle);
-
-    const deltaX = (thumstickState.xAxis ?? 0) * delta * speed;
-    const deltaY = (thumstickState.yAxis ?? 0) * delta * speed;
-    const rotatedX = deltaX * cos - deltaY * sin;
-    const rotatedZ = deltaX * sin + deltaY * cos;
-
-    // Move the player
-    ref.current.position.x += rotatedX;
-    ref.current.position.z += rotatedZ;
-  });
-  return <XROrigin ref={ref} />;
-}
 export default App;
